Added broken crosshair

+++ src/Client/Engine/Utils/Vertex.hpp
@@ -5,12 +5,12 @@
 
 
 struct GUIImageVertex {
-	GUIImageVertex(const glm::vec2& _p, GLubyte _uv){
+	GUIImageVertex(const glm::vec2& _p, GLuint _uv){
 		position = _p;
 		uv = _uv;
 	}
 	glm::vec2 position;
-	GLubyte uv = 0;
+	GLuint uv = 0;
 };
 
 struct GUITextVertex {
